+++ pkg/engine/condition/main.go
@@ -1,0 +1,9 @@
+package condition
+
+// Condition defines which condition needs to be met
+// in order to update targets based on the source output
+type Condition struct {
+	Name string
+	Kind string
+	Spec interface{}
+}
