+++ pkg/engine/source/main.go
@@ -1,0 +1,10 @@
+package source
+
+// Source defines how a value is retrieved from a specific source
+type Source struct {
+	Kind    string
+	Output  string
+	Prefix  string
+	Postfix string
+	Spec    interface{}
+}
